+++ a2a-demo/rpc-agent3/src/main/java/space/ruiwang/RpcAgent3Application.java
@@ -3,14 +3,14 @@ package space.ruiwang;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
-import space.ruiwang.agent.annotation.EnableRpcAgent;
+import space.ruiwang.agent.annotation.EnableAgent;
 import space.ruiwang.annotation.EnableAgentConsumerRpc;
 import space.ruiwang.annotation.EnableAgentProviderRpc;
 
 @SpringBootApplication
 @EnableAgentConsumerRpc
 @EnableAgentProviderRpc
-@EnableRpcAgent
+@EnableAgent
 public class RpcAgent3Application {
     public static void main(String[] args) {
         SpringApplication.run(RpcAgent3Application.class, args);