+++ rpc-framework/rpc-spring-boot-starter/src/main/java/space/ruiwang/agent/AgentRpcConsumerConfiguration.java
@@ -1,4 +1,4 @@
-package space.ruiwang;
+package space.ruiwang.agent;
 
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.context.annotation.Bean;