+++ rpc-framework/rpc-spring-boot-starter/src/main/java/space/ruiwang/agent/DefaultAgentSkillRpcService.java
@@ -1,4 +1,4 @@
-package space.ruiwang;
+package space.ruiwang.agent;
 
 import org.springframework.beans.factory.ObjectProvider;